+++ src/App.js
@@ -1,33 +1,16 @@
 import React from 'react';
-import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
+import {  Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import QuizPage from './pages/QuizPage';
 
 function App() {
   return (
-    <Router>
-      <div>
-        <nav>
-          <ul>
-            <li>
-              <Link to="/">Home</Link>
-            </li>
-            <li>
-              <Link to="/quiz">Quiz</Link>
-            </li>
-          </ul>
-        </nav>
-
-        <Switch>
-          <Route path="/quiz">
-            <QuizPage />
-          </Route>
-          <Route path="/">
-            <Home />
-          </Route>
-        </Switch>
-      </div>
-    </Router>
+    <div>
+    <Routes>
+      <Route exact path="/QuizPage" element={<QuizPage/>}/>
+      <Route exact path="/" element={<Home/>}/>
+    </Routes>
+    </div>
   );
 }
